Change unnecessary throws to returns
We should expect that the user may accidentally try to stack ships or
place ships out of bounds; in those cases we should not place that ship
and we should return that the attempt was unsuccessful, rather than
throw an error. We should do the same if the gameboard receives an
attempt to place a ship diagonally.