Record hits and missed shots properly
Instead of having the misses property of the gameboard class publicly
accessible, I added a get method for retrieving the misses as an array.
A set will add an array even if there is already another array present
containing the same elements, so I used JSON.stringify to store hits and
misses in their respective sets without duplication. This also allows us
to correctly check whether or not an attack has been received at a given
pair of coordinates.