Refactor placeShip method and its helpers
I extracted the part of the placeShip method and its helpers that
validates whether or not a new ship can be placed at the provided start
and end coordinates and placed this functionality into a new helper. Now
only the new helper, isValidShipPlacement, returns whether or not a ship
may be placed at the provided start and end coordinates, while placeShip
and its other helpers do not.